AR: Kreider Ar, version 1.3
It's always nice when a program's author does the updates, because they
usually use exactly the same environment they made the program with in the
first place. When other people do it, there are always little differences
in code style, editor, etc. that make it more difficult to extract only
the real differences between two code versions.

In Ar 1.3, the archiver gains the ability to store, display, and restore
permission flags. It drops some code from the support file (which means it
now requires Carl's version of the C library instead of Microware's), and
the LZ1 code brings in some of the changes from Greg Walker's VAX/VMS port
(though nothing that was VAX-specific).

Carl also switched his build system on 6809, from something _like_ Make
to something resembling the actual Make (possibly the "TK" Make). As such,
there aren't different versions of the Makefile any more, one file handles
both OS9 and OSK.

+++ Source/Archiver/AR/delz1.c
@@ -62,6 +62,10 @@ long     bytes;
 /*page*/
 /*
  * Read one code from the input file.  If EOF, return -1.
+ * Inputs:
+ *      infile:    Input file.
+ * Outputs:
+ *      code or -1 is returned.
  */
 
 WORD
@@ -108,14 +112,14 @@ fetch()
 
    if (w_offset + n_bits <= WSIZE)
       {
-      return((buf[word] >> ((WSIZE - n_bits) - w_offset))
+      return(((unsigned) buf[word] >> ((WSIZE - n_bits) - w_offset))
          & LowOrder(n_bits));
       }
    else
       {
       size2 = n_bits - (WSIZE - w_offset);
       return(((buf[word] << size2) |
-            (buf[word + 1] >> (WSIZE - size2)))
+            ((unsigned) buf[word + 1] >> (WSIZE - size2)))
             & LowOrder(n_bits));
       }
    }
